* refactor code
